+++ tests/lib/rules/jsx-max-props-per-line.js
@@ -60,7 +60,70 @@ ruleTester.run('jsx-max-props-per-line', rule, {
       '/>'
     ].join('\n'),
     options: [{maximum: 2}]
-  }],
+  }, {
+    code: [
+      '<App',
+      '  foo bar',
+      '  baz',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {multi: 2}}]
+  }, {
+    code: [
+      '<App',
+      '  bar',
+      '  baz',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {multi: 2, single: 1}}]
+  }, {
+    code: '<App foo baz bar />',
+    options: [{maximum: {multi: 2, single: 3}}]
+  }, {
+    code: '<App {...this.props} bar />',
+    options: [{maximum: {single: 2}}]
+  }, {
+    code: [
+      '<App',
+      '  foo bar',
+      '  baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {multi: 2, single: 1}}]
+  }, {
+    code: '<App foo baz bar />',
+    options: [{maximum: {multi: 2}}]
+  }, {
+    code: [
+      '<App',
+      '  foo bar',
+      '  baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {single: 1}}]
+  }, {
+    code: [
+      '<App foo bar',
+      '  baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {single: 2, multi: 2}}]
+  }, {
+    code: [
+      '<App foo bar',
+      '  baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: 2}]
+  }, {
+    code: [
+      '<App foo',
+      '  bar',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: 1, when: 'multiline'}]
+  }
+  ],
 
   invalid: [{
     code: '<App foo bar baz />;',
@@ -262,6 +325,122 @@ ruleTester.run('jsx-max-props-per-line', rule, {
       '/>'
     ].join('\n'),
     options: [{maximum: 2}],
+    errors: [{
+      messageId: 'newLine',
+      data: {prop: 'baz'}
+    }]
+  },
+  {
+    code: '<App foo bar baz />',
+    output: [
+      '<App foo',
+      'bar',
+      'baz />'
+    ].join('\n'),
+    options: [{maximum: {single: 1, multi: 1}}],
+    errors: [{
+      messageId: 'newLine',
+      data: {prop: 'bar'}
+    }]
+  }, {
+    code: [
+      '<App',
+      '  foo bar baz',
+      '/>'
+    ].join('\n'),
+    output: [
+      '<App',
+      '  foo',
+      'bar',
+      'baz',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {single: 1, multi: 1}}],
+    errors: [{
+      messageId: 'newLine',
+      data: {prop: 'bar'}
+    }]
+  }, {
+    code: [
+      '<App foo',
+      '  bar baz',
+      '/>'
+    ].join('\n'),
+    output: [
+      '<App foo',
+      '  bar',
+      'baz',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {single: 1, multi: 1}}],
+    errors: [{
+      messageId: 'newLine',
+      data: {prop: 'baz'}
+    }]
+  }, {
+    code: [
+      '<App foo bar',
+      '  bar baz bor',
+      '/>'
+    ].join('\n'),
+    output: [
+      '<App foo bar',
+      '  bar baz',
+      'bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {single: 1, multi: 2}}],
+    errors: [
+      {
+        messageId: 'newLine',
+        data: {prop: 'bor'}
+      }]
+  }, {
+    code: '<App foo bar baz bor />',
+    output: [
+      '<App foo bar baz',
+      'bor />'
+    ].join('\n'),
+    options: [{maximum: {single: 3, multi: 2}}],
+    errors: [
+      {
+        messageId: 'newLine',
+        data: {prop: 'bor'}
+      }]
+  }, {
+    code: [
+      '<App',
+      '  foo={{',
+      '  }} bar baz bor',
+      '/>'
+    ].join('\n'),
+    output: [
+      '<App',
+      '  foo={{',
+      '  }} bar',
+      'baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {multi: 2}}],
+    errors: [{
+      messageId: 'newLine',
+      data: {prop: 'baz'}
+    }]
+  }, {
+    code: [
+      '<App boz fuz',
+      '  foo={{',
+      '  }} bar baz bor',
+      '/>'
+    ].join('\n'),
+    output: [
+      '<App boz fuz',
+      '  foo={{',
+      '  }} bar',
+      'baz bor',
+      '/>'
+    ].join('\n'),
+    options: [{maximum: {multi: 2, single: 1}}],
     errors: [{
       messageId: 'newLine',
       data: {prop: 'baz'}