+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Added
 * add [`no-namespace`] rule ([#2640] @yacinehmito @ljharb)
+* [`jsx-max-props-per-line`]: add `single` and `multi` options ([#3078] @SIL0RAK)
 
 ### Fixed
 * [`display-name`]: Get rid of false position on component detection ([#2759] @iiison)
@@ -14,6 +15,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 ### Changed
 * [`no-access-state-in-setstate`]: passing test for “don't error if it's not a React Component” ([#1873] @kentcdodds)
 
+[#3078]: https://github.com/yannickcr/eslint-plugin-react/pull/3078
 [#2640]: https://github.com/yannickcr/eslint-plugin-react/pull/2640
 [#2759]: https://github.com/yannickcr/eslint-plugin-react/pull/2759
 [#1873]: https://github.com/yannickcr/eslint-plugin-react/pull/1873